Melodie Cell Wide Version
Tried to fix Melodie's cell textures, managed to do most except for the emergency lights as it seems to lose color on the ones at the right side.

Made what i did as a Prefab in "Melodie Cell Wide Version" folders so it can be dragged in and out easily.